Move `CaptureLayer` to separate module
